Add files via upload
I utilized Microsoft Copilot to help me create the test suite. I input my code and asked it to create a test suite with 80% coverage. I spent a long time trying to have it generate tests that would cover the entirety of the code and pass, but I struggled and 33% was the best I could get.  It was a lot of trial and error.